Refactor semantic model - move things around
