+++ crates/semantic/src/corelib.rs
@@ -3,7 +3,7 @@ use filesystem::ids::CrateLongId;
 use syntax::token::TokenKind;
 
 use crate::db::SemanticGroup;
-use crate::{ConcreteType, TypeId, TypeLongId};
+use crate::{semantic, TypeId};
 
 pub fn core_module(db: &dyn SemanticGroup) -> ModuleId {
     let core_crate = db.intern_crate(CrateLongId("core".into()));
@@ -15,11 +15,14 @@ pub fn core_felt_ty(db: &dyn SemanticGroup) -> TypeId {
     // This should not fail if the corelib is present.
     let generic_type =
         db.module_item_by_name(core_module, "felt".into()).and_then(GenericTypeId::from).unwrap();
-    db.intern_type(TypeLongId::Concrete(ConcreteType { generic_type, generic_args: vec![] }))
+    db.intern_type(semantic::TypeLongId::Concrete(semantic::ConcreteType {
+        generic_type,
+        generic_args: vec![],
+    }))
 }
 
 pub fn unit_ty(db: &dyn SemanticGroup) -> TypeId {
-    db.intern_type(TypeLongId::Tuple(vec![]))
+    db.intern_type(semantic::TypeLongId::Tuple(vec![]))
 }
 
 pub fn core_binary_operator(
