+++ crates/semantic/src/items/mod.rs
@@ -1,0 +1,3 @@
+pub mod free_function;
+pub mod functions;
+pub mod strct;
