+++ crates/semantic/src/lib.rs
@@ -7,8 +7,10 @@ pub mod db;
 pub mod diagnostic;
 pub mod expr;
 pub mod ids;
+pub mod items;
 mod resolve_item;
 mod semantic;
+pub mod types;
 
 pub use diagnostic::SemanticDiagnostic;
 pub use ids::*;
